[python] fix: annotate the package so it type-checks cleanly
`mypy` reported 44 errors across 25 files. None of them was a runtime bug —
every one was an annotation that did not describe what the code already did —
so no behavior changes here.

Most were implicit `Optional`: eleven parameters declared `str`, `bool`, `list`
or `datetime` while defaulting to `None`, which PEP 484 has not allowed since
mypy turned `no_implicit_optional` on by default. `_jsParseInt` was declared
`-> float` but only ever returns an `int`, and `safeParseInt` returns its
result as its own `int`. `arrUnique` bound `seen` to a set of JSON keys in one
branch and a list of items in the other, so the keys are now `seenKeys`.
Empty containers, the `_UNITS` table and the `_cipher` mode variable say what
they hold. `fetchData` reads `host` and `onError` into locals, which cannot be
narrowed across two `opt.get` calls and is one lookup fewer besides.

Two needed a word about why they are safe rather than a wider type.
`_decimalShift` adds to the exponent of a `Decimal`, which is a string for NaN
and infinity; every caller checks `is_finite()` first, and that guarantee is now
written down. `arrTo1dArray` walks tuples as well as lists and hands them to
`is2dArray`, which reads any sequence but is annotated for the public contract.

`mypy` is a dev dependency and configured in `pyproject.toml`, so `mypy` from
the package root reproduces the clean run.

+++ packages/python/qsu/format/safeParseInt.py
@@ -1,7 +1,7 @@
 from typing import Any
 
 
-def _jsParseInt(s: str, radix: int) -> float:
+def _jsParseInt(s: str, radix: int) -> int:
 	# Mirror JavaScript's parseInt leniency: parse the leading valid prefix,
 	# ignore the rest, and return NaN-like behavior via ValueError to the caller.
 	s = s.strip()
